+++ canopen_analyzer/dox/documentation/html/search/all_f.js
@@ -1,16 +1,16 @@
 var searchData=
 [
-  ['rate_5fmisc_154',['rate_misc',['../classdisplay__gui_1_1CANopenMainWindow.html#a692dccfb286da085c50701bb0655e44a',1,'display_gui::CANopenMainWindow']]],
-  ['rate_5fpdo_155',['rate_pdo',['../classdisplay__gui_1_1CANopenMainWindow.html#ad18bd90639b22805e6d23ba2ec4e4334',1,'display_gui::CANopenMainWindow']]],
-  ['rate_5fsdo_156',['rate_sdo',['../classdisplay__gui_1_1CANopenMainWindow.html#af9aade33e4ca0f15067d6005207620b9',1,'display_gui::CANopenMainWindow']]],
-  ['rates_5fstats_157',['rates_stats',['../classbus__stats_1_1bus__stats_1_1rates__stats.html',1,'bus_stats::bus_stats']]],
-  ['raw_5fframe_158',['raw_frame',['../classcanopen__sniffer_1_1canopen__sniffer.html#a796de4da9ed5fad20bff2ba1ea3b7051',1,'canopen_sniffer.canopen_sniffer.raw_frame()'],['../classprocess__frames_1_1process__frames.html#a796de4da9ed5fad20bff2ba1ea3b7051',1,'process_frames.process_frames.raw_frame()']]],
-  ['readme_2emd_159',['README.md',['../README_8md.html',1,'']]],
-  ['refresh_5finterval_160',['refresh_interval',['../classdisplay__tui_1_1display__tui.html#a5f920440f90c3d64d4ef76fd55e2a589',1,'display_tui::display_tui']]],
-  ['render_5ftables_161',['render_tables',['../classdisplay__cli_1_1display__cli.html#a1602e5a067d7cbe99e22f10101f9e381',1,'display_cli::display_cli']]],
-  ['reset_162',['reset',['../classbus__stats_1_1bus__stats.html#a51829b63adb24ac48d350dee60181002',1,'bus_stats::bus_stats']]],
-  ['resizeevent_163',['resizeEvent',['../classdisplay__gui_1_1MultiRateLineWidget.html#a0815964d2d7e4debf9b3e7cff7840436',1,'display_gui::MultiRateLineWidget']]],
-  ['root_5flogger_164',['root_logger',['../namespaceanalyzer__defs.html#af2fe9aec0484a8d331ffcbc182d64af3',1,'analyzer_defs']]],
-  ['run_165',['run',['../classcanopen__sniffer_1_1canopen__sniffer.html#ad22709b2e67308af35f55680d5a026e0',1,'canopen_sniffer.canopen_sniffer.run()'],['../classdisplay__cli_1_1display__cli.html#ad22709b2e67308af35f55680d5a026e0',1,'display_cli.display_cli.run()'],['../classdisplay__gui_1_1GUIUpdateWorker.html#ad22709b2e67308af35f55680d5a026e0',1,'display_gui.GUIUpdateWorker.run()'],['../classprocess__frames_1_1process__frames.html#ad22709b2e67308af35f55680d5a026e0',1,'process_frames.process_frames.run()']]],
-  ['run_5ftextual_166',['run_textual',['../classdisplay__tui_1_1display__tui.html#a06d308c260a5e945f80724fcd934a8c7',1,'display_tui::display_tui']]]
+  ['rate_5fmisc_156',['rate_misc',['../classdisplay__gui_1_1CANopenMainWindow.html#a692dccfb286da085c50701bb0655e44a',1,'display_gui::CANopenMainWindow']]],
+  ['rate_5fpdo_157',['rate_pdo',['../classdisplay__gui_1_1CANopenMainWindow.html#ad18bd90639b22805e6d23ba2ec4e4334',1,'display_gui::CANopenMainWindow']]],
+  ['rate_5fsdo_158',['rate_sdo',['../classdisplay__gui_1_1CANopenMainWindow.html#af9aade33e4ca0f15067d6005207620b9',1,'display_gui::CANopenMainWindow']]],
+  ['rates_5fstats_159',['rates_stats',['../classbus__stats_1_1bus__stats_1_1rates__stats.html',1,'bus_stats::bus_stats']]],
+  ['raw_5fframe_160',['raw_frame',['../classcanopen__sniffer_1_1canopen__sniffer.html#a796de4da9ed5fad20bff2ba1ea3b7051',1,'canopen_sniffer.canopen_sniffer.raw_frame()'],['../classprocess__frames_1_1process__frames.html#a796de4da9ed5fad20bff2ba1ea3b7051',1,'process_frames.process_frames.raw_frame()']]],
+  ['readme_2emd_161',['README.md',['../README_8md.html',1,'']]],
+  ['refresh_5finterval_162',['refresh_interval',['../classdisplay__tui_1_1display__tui.html#a5f920440f90c3d64d4ef76fd55e2a589',1,'display_tui::display_tui']]],
+  ['render_5ftables_163',['render_tables',['../classdisplay__cli_1_1display__cli.html#a1602e5a067d7cbe99e22f10101f9e381',1,'display_cli::display_cli']]],
+  ['reset_164',['reset',['../classbus__stats_1_1bus__stats.html#a51829b63adb24ac48d350dee60181002',1,'bus_stats::bus_stats']]],
+  ['resizeevent_165',['resizeEvent',['../classdisplay__gui_1_1MultiRateLineWidget.html#a0815964d2d7e4debf9b3e7cff7840436',1,'display_gui::MultiRateLineWidget']]],
+  ['root_5flogger_166',['root_logger',['../namespaceanalyzer__defs.html#af2fe9aec0484a8d331ffcbc182d64af3',1,'analyzer_defs']]],
+  ['run_167',['run',['../classcanopen__sniffer_1_1canopen__sniffer.html#ad22709b2e67308af35f55680d5a026e0',1,'canopen_sniffer.canopen_sniffer.run()'],['../classdisplay__cli_1_1display__cli.html#ad22709b2e67308af35f55680d5a026e0',1,'display_cli.display_cli.run()'],['../classdisplay__gui_1_1GUIUpdateWorker.html#ad22709b2e67308af35f55680d5a026e0',1,'display_gui.GUIUpdateWorker.run()'],['../classprocess__frames_1_1process__frames.html#ad22709b2e67308af35f55680d5a026e0',1,'process_frames.process_frames.run()']]],
+  ['run_5ftextual_168',['run_textual',['../classdisplay__tui_1_1display__tui.html#a06d308c260a5e945f80724fcd934a8c7',1,'display_tui::display_tui']]]
 ];