+++ canopen_analyzer/dox/documentation/html/search/classes_4.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['frame_5fcount_210',['frame_count',['../classbus__stats_1_1bus__stats_1_1frame__count.html',1,'bus_stats::bus_stats']]],
-  ['frame_5ftype_211',['frame_type',['../classanalyzer__defs_1_1frame__type.html',1,'analyzer_defs']]]
+  ['frame_5fcount_213',['frame_count',['../classbus__stats_1_1bus__stats_1_1frame__count.html',1,'bus_stats::bus_stats']]],
+  ['frame_5ftype_214',['frame_type',['../classanalyzer__defs_1_1frame__type.html',1,'analyzer_defs']]]
 ];