+++ CHANGELOG.md
@@ -18,6 +18,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Bus Active/ Idle detection.
+- Reset bus stats on nodes becoming inactive.
+- Added active Node Ids to CLI, TUI and GUI.
 
 ## [v0.16.0] - 2025-12-22